pyproject.toml
[tool.poetry]
name = "alnoda_wrk"
version = "0.3.2"
description = "A tool to build Alnoda workspaces"
authors = ["bluxmit <bluxmit@gmail.com>"]
license = "GNU Affero General Public License v3.0"
readme = "README.md"

[tool.poetry.scripts]
"alnoda-wrk" = "alnoda_wrk.main:app"
"wrk" = "alnoda_wrk.main:app"

[tool.poetry.dependencies]
python = "^3.8"
Cerberus = "^1.3" 
jinja2 = "^3.0"
PyYAML = "^6.0"
typer = {version = "^0.6", extras = ["all"]}
pyfiglet = "^0.8.post1"
pyTermTk = "^0.10.22a0" 
requests = "^2.28.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
